backend/requirements.txt
django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.1
psycopg2-binary==2.9.10
django-environ==0.12.0
dj-database-url==3.0.1
redis==6.4.0
django-redis==6.0.0
celery==5.5.3
google-auth==2.40.3
razorpay==2.0.0
sendgrid==6.12.5
twilio==9.8.1
fcm-django==2.3.1
drf-spectacular==0.28.0
drf-nested-routers==0.95.0
gunicorn==23.0.0
